Modules: Restructured parsing code.
Module can now return a json_t object from rpi_module_get_value_t, and it will get parsed in REST api urls.

+++ src/include/rpi_memory.h
@@ -4,12 +4,12 @@
 #define RPI_MEMORY_H
 
 typedef struct {
-    int total;
-    int free;
-    int buffers;
-    int cached;
-    int swap_total;
-    int swap_free;
+    long total;
+    long free;
+    long buffers;
+    long cached;
+    long swap_total;
+    long swap_free;
 } rpi_memory_buffer_t;
 
 void rpi_memory_init(void);
